refactor: departure view with separate modules for sub-views
