+++ frontend/src/views/departureEmptyStateView.tsx
@@ -1,0 +1,3 @@
+export function DepartureEmptyStateView() {
+    return <p className="text-sm text-slate-600">No upcoming departures</p>;
+}
