+++ frontend/src/views/departureView.tsx
@@ -1,15 +1,14 @@
+import { Suspense } from "../components/Suspense";
 import { Departure } from "../types/sl";
 import { DelaySummary } from "../types/historicalDelay";
-import { Suspense } from "../components/Suspense";
-import Button from "@mui/material/Button";
-import StarIcon from "@mui/icons-material/Star";
-import StarBorderIcon from "@mui/icons-material/StarBorder";
-import { DepartureDetails } from "../components/DepartureDetails";
-import { DepartureList } from "../components/DepartureList";
 import { DatePreset } from "../types/departureDelay";
+import { DepartureHeaderView } from "./departureHeaderView";
+import { DepartureEmptyStateView } from "./departureEmptyStateView";
+import { DepartureList } from "../components/DepartureList";
+import { DepartureDetails } from "../components/DepartureDetails";
 
 export type DepartureViewProps = {
-    departures: Departure[];
+    upcomingDepartures: Departure[];
     selectedDeparture: Departure | null;
     selectedSiteName: string;
     onClose: () => void;
@@ -30,7 +29,7 @@ export type DepartureViewProps = {
 };
 
 export function DepartureView({
-    departures,
+    upcomingDepartures,
     selectedDeparture,
     selectedSiteName,
     onClose,
@@ -49,82 +48,15 @@ export function DepartureView({
     isUserLoggedIn,
     onToggleFavoriteStop,
 }: DepartureViewProps) {
-    const nonUpcomingStates = new Set([
-        "DEPARTED",
-        "PASSED",
-        "MISSED",
-        "ASSUMEDDEPARTED",
-        "CANCELLED",
-        "INHIBITED",
-        "NOTCALLED",
-        "REPLACED",
-    ]);
-
-    function getDepartureTimestampCB(departure: Departure) {
-        const candidateTime = departure.expected ?? departure.scheduled;
-        const parsedTime = Date.parse(candidateTime);
-
-        return Number.isNaN(parsedTime) ? null : parsedTime;
-    }
-
-    function isUpcomingDepartureCB(departure: Departure) {
-        const timestamp = getDepartureTimestampCB(departure);
-
-        return timestamp !== null && !nonUpcomingStates.has(departure.state);
-    }
-
-    const upcomingDepartures = departures.filter(isUpcomingDepartureCB).sort(compareDeparturesCB);
-
-    function compareDeparturesCB(a: Departure, b: Departure) {
-        const aTime = getDepartureTimestampCB(a);
-        const bTime = getDepartureTimestampCB(b);
-
-        if (aTime === null && bTime === null) {
-            return 0;
-        }
-        if (aTime === null) {
-            return 1;
-        }
-        if (bTime === null) {
-            return -1;
-        }
-        return aTime - bTime;
-    }
-
     return (
         <div className="flex flex-col gap-2">
-            <div className="flex items-center justify-between gap-2">
-                <h3 className="text-sm font-medium text-slate-800">{selectedSiteName}</h3>
-                <div className="flex items-center gap-1">
-                    <Button
-                        variant="text"
-                        size="small"
-                        onClick={onToggleFavoriteStop}
-                        startIcon={isFavoriteStop ? <StarIcon /> : <StarBorderIcon />}
-                        aria-label={
-                            isUserLoggedIn
-                                ? isFavoriteStop
-                                    ? "Remove stop from favorites"
-                                    : "Add stop to favorites"
-                                : "Log in to save favorites"
-                        }
-                    >
-                        {isUserLoggedIn
-                            ? isFavoriteStop
-                                ? "Unfavorite"
-                                : "Favorite"
-                            : "Log in to favorite"}
-                    </Button>
-                    <Button
-                        variant="text"
-                        size="small"
-                        onClick={onClose}
-                        aria-label="Close departures view"
-                    >
-                        Close
-                    </Button>
-                </div>
-            </div>
+            <DepartureHeaderView
+                selectedSiteName={selectedSiteName}
+                isFavoriteStop={isFavoriteStop}
+                isUserLoggedIn={isUserLoggedIn}
+                onToggleFavoriteStop={onToggleFavoriteStop}
+                onClose={onClose}
+            />
             {isLoading ? (
                 <Suspense message="Loading departures..." />
             ) : selectedDeparture ? (
@@ -146,7 +78,7 @@ export function DepartureView({
                     onSelectDeparture={onSelectDeparture}
                 />
             ) : (
-                <p className="text-sm text-slate-600">No upcoming departures</p>
+                <DepartureEmptyStateView />
             )}
         </div>
     );
